move all staking tasks to a different file
